doc i700101 record initial architecture decisions
The backend of this program is obviously java. The frontend at this
time is a javascript rendered canvas. The program will somewhat
resemble the actor model, whereby different processes produce and
consume messages that arrive via a concurrent queue.

These are now explicit for someone perusing the documents of this
project before diving in deeply to see how it deals with these issues.