feat i20220823 rudimentary showing map
Connects to a websocket page that shows the nodes of the atc map
as ascii characters. Adds caches for airplanes and the map.
Adds CoalesceBlobs to print otherwise ignorable atc responses
that are canonically satisfied by the arrival of an atc event like
game started. This shows the node that the airplanes arrive on,
but does not update when the flight plan changes.
Adds a visualization of the (message) queues connecting various
parts of haokkoz.