feat i20220914 canvas able to end session
Makeshift means of quitting the haokkoz session via the canvas.
Removes the timed end in HaoStarter. Web page declares the default
key to quit is p. Partially updates module graphviz file with this
queue edge. Removes explicit quit methods from interfaces in favor
of extending Quittable. The canvas endpoint is handling the mapping
for this moment, but that should move to BoardMapper or a different
mapper or the adapter, at minimum.