brk chg i20220914 startup requires view, only via queues
Sending a start message to atc requires a view to connect.
HaoStarter.GameRunner (almost) entirely handles the initialization
by requesting the game version (to ensure it is compatible),
starting the event stream, and starting the game once it has
notified the event stream and a view is connected.
This is entirely through queues, rather than calling through
the interface.

This stops the server either when the atc game ends or the
view requests a quit. When the view can start a new game,
then we can sever the queue that stops the server when
the atc game ends.
While this uses Move as a mhc 'view connected' message, it
may behoove hao to use a different message, perhaps creating one.