feat i20220823 able to zone nodes
Uses the MapServiceEndpoint to request three node points. It treats
these as a triangle to estimate rectangles centered on each node.
It then returns a map with the estimated rectangles, aka zones, for
all nodes of the map. Soon, haokkoz will use these zones to estimate
which node a given airplane is near.

Centralizes the milliseconds to sleep. Reduces this duration, given
that some of these messages end up taking six or more trips through
queues, which could build up if each is 1/5 of a second.

Avoids sending a start game request until the view is listening
and the event stream has started.