fix i20220823 updates airplane node from airplane moved event
KnowsAirplanes now offers a means of updating the airplane node
with a KnowsMap. KnowsMap offers a means of translating a point
to a node. So, CanvasAdapter, before building the mhc version of
a map, asks to update the map nodes of all airplanes. If some node
changed and the repaint reason was a node changing (but actually a
pixelwise change of an airplane), then it will repaint. Other
repaint reasons succeed as normal.