fix i20220823 consistent airplane character, flight plan color
To assist with diagnosing the airplanes, this uses a consistent
character, independent of the entire id. While it would be nice
to use the number directly on the character space, it is interspersed
with unhelpful characters, so AtcAirplane uses a safe alphabet
and loops over it.
This also darkens the nodes as the distance in the flight plan
from the initial node increases.