feat i20220803 consistent id, old flight plan in black
Chooses a consistent airplane char on canvas based on the id.
Wraps around within the ascii range, to avoid unprintable chars.
I regard it acceptable to reuse the same char if there are 90
airplanes in flight.

Draws the old flight plan nodes in black, though we overwrite
these with newer entities.