feat i20220815 able to direct planes away from arrival
Able to request a flight plan that moves straight away from
the nearest map edge. Anoints AirplaneDispatch with this
responsibility, as it has access to the airplane information
needed, already.

Sends airplane moved events to MapDispatch. While this is
important to airplanes, the real information is whether the
new position has crossed the boundary to a new node. Hence,
the service that can tell us which node it is near will
augment the message with the desired information and forward
it to AirplaneDispatch.

Unfortunately, this indicates that the current design of
iterating over all the rectangles does not accomplish anything.
While flipping the map so rectangles are the keys is simple,
only a better structure would remove the need for the same
iteration. Very likely, this will change to a geometric
calculation instead, given the regularity of the landmarks.