fix i20220823 keep aircraft char visible
This restricts the aircraft character for the canvas within the
visible ascii range. While this could continue to increment
up to unicode with diacritics, invisible characters follow
the plain ascii range. While this means that there will be
repeats if there are more than 90 airplanes on the screen,
that's a pretty busy map, so it won't really matter.