fix i20220823 seems to update airplane positions regularly
This now only updates the airplane closest node when the position
is closer to a different node, rather than updating it based
on the arrival at a known location (ex the start/end of a
flight path). Using an experientially calculated node distance.
Perhaps I could get things to line up, but this required some
violence given the fact that I swapped visual lat/long meanings
and had negated them based on insufficient evidence, and so on.

Removes the unnecessary repetition of interrupted exception.
I already know everything is shutting down.